+++ config/database.go
@@ -1,0 +1,5 @@
+package config
+
+const (
+	DbName = "./nyaa.db"
+)
