+++ config/env.go
@@ -1,0 +1,8 @@
+package config
+
+// Constants for environment.
+const (
+	// DEVELOPMENT | TEST | PRODUCTION
+	Environment = "DEVELOPMENT"
+	// Environment            = "PRODUCTION"
+)
