+++ config/navigation.go
@@ -1,0 +1,6 @@
+package config
+
+// Constants for pagination.
+const (
+	TorrentsPerPage  = 50
+)
