+++ config/sorting.go
@@ -1,0 +1,7 @@
+package config
+
+// Constants for ordering models.
+const (
+	TorrentOrder  = "torrent_id"
+	TorrentSort  = "DESC"
+)
