+++ util/safe.go
@@ -1,0 +1,7 @@
+package util
+
+import 	"html/template"
+
+func Safe(s string) template.URL {
+	return template.URL(s)
+}
